Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -62,7 +62,7 @@ def get_average_age_for_season(data, season)
     if (k==season)
       n=0
       v.each do |h|
-        a.push(data[k][n]["age"].to_i)
+        a.push(data[k][n]["age"].to_f)
         n+=1
       end
     end
